# perf(ext/web): optimize DOMException